Create Source code
This is a Java source code that I developed that allows the user to select up to 3 snacks at a time and prints a receipt of the transaction. There is also a way to shut down the vending machine with 2FA. You must use the code 999 at the beginning and the password 'COOKIES' following that 999 code.